stress test updates
git-svn-id: svn://svn.caucho.com/resin/trunk@6075 9c94448d-38f1-0310-a231-d98308ff1ebf

+++ modules/resin/src/com/caucho/db/table/Column.java
@@ -424,7 +424,7 @@ abstract public class Column {
    * @param rowOffset the offset of the row in the block
    * @param value the value to store
    */
-  void setDate(Transaction xa, byte []block, int rowOffset, double value)
+  void setDate(Transaction xa, byte []block, int rowOffset, long value)
     throws SQLException
   {
     throw new UnsupportedOperationException();
